Add files via upload
# Pizza Menu

This project is a simple pizza menu application built using React. The application displays a list of pizzas with their names, descriptions, and prices. It is styled using CSS to create a visually appealing layout. This project is a basic example of how to use React for creating dynamic web applications.

## Getting Started

### Prerequisites

- Node.js
- npm

### Installation

1. Clone the repository:
   ```bash
   git clone https://github.com/your-username/pizza-menu.git
   cd pizza-menu